Refactor variational module
