+++ pymc3/variational/__init__.py
@@ -1,2 +1,3 @@
-from .advi import advi, sample_vp
+from .advi import advi
 from .advi_minibatch import advi_minibatch
+from .sample_vp import sample_vp
